# specs/001-ecommerce-site/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Fish Market API
  version: 1.0.0
  description: API contract for the full-stack demo storefront.
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged-in user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  schemas:
...
    User:
      type: object
      required: [id, displayName, email, role]
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
...
    RegisterRequest:
      type: object
      required: [displayName, email, password]
      properties:
        displayName:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
...